fix(tests): update test cases for calc_fee
