+++ wallet/client.go
@@ -23,6 +23,7 @@ type grpcClient struct {
 
 	blockchainClient  pactus.BlockchainClient
 	transactionClient pactus.TransactionClient
+	utilityClient     pactus.UtilityClient
 }
 
 func newGRPCClient(rpcEndpoint string) (*grpcClient, error) {
@@ -101,7 +102,7 @@ func (c *grpcClient) getTransaction(id tx.ID) (*pactus.GetTransactionResponse, e
 }
 
 func (c *grpcClient) getFee(amount int64, payloadType payload.Type) (int64, error) {
-	res, err := c.transactionClient.CalculateFee(c.ctx,
+	res, err := c.utilityClient.CalculateFee(c.ctx,
 		&pactus.CalculateFeeRequest{Amount: amount, PayloadType: pactus.PayloadType(payloadType)})
 	if err != nil {
 		return 0, err
