+++ www/grpc/transaction_test.go
@@ -117,23 +117,6 @@ func TestSendRawTransaction(t *testing.T) {
 	td.StopServer()
 }
 
-func TestGetCalculateFee(t *testing.T) {
-	td := setup(t, nil)
-	conn, client := td.transactionClient(t)
-
-	amount := td.RandAmount()
-	res, err := client.CalculateFee(context.Background(),
-		&pactus.CalculateFeeRequest{
-			Amount:      amount,
-			PayloadType: pactus.PayloadType_TRANSFER_PAYLOAD,
-		})
-	assert.NoError(t, err)
-	assert.Equal(t, amount/10000, res.Fee)
-
-	assert.Nil(t, conn.Close(), "Error closing connection")
-	td.StopServer()
-}
-
 func TestGetRawTransaction(t *testing.T) {
 	td := setup(t, nil)
 	conn, client := td.transactionClient(t)
@@ -221,38 +204,6 @@ func TestGetRawTransaction(t *testing.T) {
 		assert.Equal(t, amount, decodedTrx.Payload().Value())
 		assert.Equal(t, expectedLockTime, decodedTrx.LockTime())
 		assert.Equal(t, expectedFee, decodedTrx.Fee())
-	})
-
-	assert.Nil(t, conn.Close(), "Error closing connection")
-	td.StopServer()
-}
-
-func TestCalculateFee(t *testing.T) {
-	td := setup(t, nil)
-	conn, client := td.transactionClient(t)
-
-	t.Run("Not fixed amount", func(t *testing.T) {
-		amount := td.RandAmount()
-		res, err := client.CalculateFee(context.Background(),
-			&pactus.CalculateFeeRequest{
-				Amount:      amount,
-				PayloadType: pactus.PayloadType_TRANSFER_PAYLOAD,
-				FixedAmount: false,
-			})
-		assert.NoError(t, err)
-		assert.Equal(t, res.Amount, amount)
-	})
-
-	t.Run("Fixed amount", func(t *testing.T) {
-		amount := td.RandAmount()
-		res, err := client.CalculateFee(context.Background(),
-			&pactus.CalculateFeeRequest{
-				Amount:      amount,
-				PayloadType: pactus.PayloadType_TRANSFER_PAYLOAD,
-				FixedAmount: true,
-			})
-		assert.NoError(t, err)
-		assert.LessOrEqual(t, res.Amount+res.Fee, amount)
 	})
 
 	assert.Nil(t, conn.Close(), "Error closing connection")
